pyproject.toml
[tool.poetry]
name = "testtask"
version = "0.1.0"
description = "poetry req for app"
authors = ["Eugene"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.2"  # Use the latest FastAPI version
uvicorn = "^0.31.1"
sqlalchemy = "^2.0.0"
redis = "^5.0.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"  # For JWT handling
pydantic-settings = "^2.0.0" 
pydantic = "^2.0.0"
psycopg2 = ">1.0"  


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
